Move functions out of spdm_common_lib
Functions that are used exclusively by either the Requester or Responder are moved to their respective libraries. If a function is used in only one .c file then it is moved to that file and made static. If a function is used by multiple .c files it is moved to libspdm_*_common.c.
