+++ library/spdm_requester_lib/libspdm_req_finish.c
@@ -22,6 +22,277 @@ typedef struct {
 #pragma pack()
 
 /**
+ * This function verifies the finish HMAC based upon TH.
+ *
+ * @param  spdm_context                  A pointer to the SPDM context.
+ * @param  session_info                  The session info of an SPDM session.
+ * @param  hmac_data                     The HMAC data buffer.
+ * @param  hmac_data_size                 size in bytes of the HMAC data buffer.
+ *
+ * @retval true  HMAC verification pass.
+ * @retval false HMAC verification fail.
+ **/
+bool libspdm_verify_finish_rsp_hmac(libspdm_context_t *spdm_context,
+                                    libspdm_session_info_t *session_info,
+                                    const void *hmac_data, size_t hmac_data_size)
+{
+    size_t hash_size;
+    uint8_t calc_hmac_data[LIBSPDM_MAX_HASH_SIZE];
+    bool result;
+#if LIBSPDM_RECORD_TRANSCRIPT_DATA_SUPPORT
+    uint8_t *cert_chain_buffer;
+    size_t cert_chain_buffer_size;
+    uint8_t *mut_cert_chain_buffer;
+    size_t mut_cert_chain_buffer_size;
+    uint8_t th_curr_data[LIBSPDM_MAX_MESSAGE_BUFFER_SIZE];
+    size_t th_curr_data_size;
+    uint8_t hash_data[LIBSPDM_MAX_HASH_SIZE];
+#endif
+
+    hash_size = libspdm_get_hash_size(spdm_context->connection_info.algorithm.base_hash_algo);
+    LIBSPDM_ASSERT(hash_size == hmac_data_size);
+
+#if LIBSPDM_RECORD_TRANSCRIPT_DATA_SUPPORT
+    result = libspdm_get_peer_cert_chain_buffer(
+        spdm_context, (const void **)&cert_chain_buffer, &cert_chain_buffer_size);
+    if (!result) {
+        return false;
+    }
+
+    if (session_info->mut_auth_requested) {
+        result = libspdm_get_local_cert_chain_buffer(
+            spdm_context, (const void **)&mut_cert_chain_buffer,
+            &mut_cert_chain_buffer_size);
+        if (!result) {
+            return false;
+        }
+    } else {
+        mut_cert_chain_buffer = NULL;
+        mut_cert_chain_buffer_size = 0;
+    }
+    th_curr_data_size = sizeof(th_curr_data);
+    result = libspdm_calculate_th_for_finish(
+        spdm_context, session_info, cert_chain_buffer,
+        cert_chain_buffer_size, mut_cert_chain_buffer,
+        mut_cert_chain_buffer_size, &th_curr_data_size, th_curr_data);
+    if (!result) {
+        return false;
+    }
+
+    result = libspdm_hash_all (spdm_context->connection_info.algorithm.base_hash_algo,
+                               th_curr_data, th_curr_data_size, hash_data);
+    if (!result) {
+        return false;
+    }
+
+    result = libspdm_hmac_all_with_response_finished_key(
+        session_info->secured_message_context, hash_data,
+        hash_size, calc_hmac_data);
+    if (!result) {
+        return false;
+    }
+#else
+    result = libspdm_calculate_th_hmac_for_finish_rsp(
+        spdm_context, session_info, &hash_size, calc_hmac_data);
+    if (!result) {
+        return false;
+    }
+#endif
+    LIBSPDM_DEBUG((LIBSPDM_DEBUG_INFO, "th_curr hmac - "));
+    libspdm_internal_dump_data(calc_hmac_data, hash_size);
+    LIBSPDM_DEBUG((LIBSPDM_DEBUG_INFO, "\n"));
+
+    if (libspdm_const_compare_mem(calc_hmac_data, hmac_data, hash_size) != 0) {
+        LIBSPDM_DEBUG((LIBSPDM_DEBUG_INFO, "!!! verify_finish_rsp_hmac - FAIL !!!\n"));
+        return false;
+    }
+    LIBSPDM_DEBUG((LIBSPDM_DEBUG_INFO, "!!! verify_finish_rsp_hmac - PASS !!!\n"));
+
+    return true;
+}
+
+/**
+ * This function generates the finish HMAC based upon TH.
+ *
+ * @param  spdm_context                  A pointer to the SPDM context.
+ * @param  session_info                  The session info of an SPDM session.
+ * @param  hmac                         The buffer to store the finish HMAC.
+ *
+ * @retval true  finish HMAC is generated.
+ * @retval false finish HMAC is not generated.
+ **/
+static bool libspdm_generate_finish_req_hmac(libspdm_context_t *spdm_context,
+                                             libspdm_session_info_t *session_info,
+                                             void *hmac)
+{
+    size_t hash_size;
+    uint8_t calc_hmac_data[LIBSPDM_MAX_HASH_SIZE];
+    bool result;
+#if LIBSPDM_RECORD_TRANSCRIPT_DATA_SUPPORT
+    uint8_t *cert_chain_buffer;
+    size_t cert_chain_buffer_size;
+    uint8_t *mut_cert_chain_buffer;
+    size_t mut_cert_chain_buffer_size;
+    uint8_t th_curr_data[LIBSPDM_MAX_MESSAGE_BUFFER_SIZE];
+    size_t th_curr_data_size;
+    uint8_t hash_data[LIBSPDM_MAX_HASH_SIZE];
+#endif
+
+    hash_size = libspdm_get_hash_size(spdm_context->connection_info.algorithm.base_hash_algo);
+
+#if LIBSPDM_RECORD_TRANSCRIPT_DATA_SUPPORT
+    result = libspdm_get_peer_cert_chain_buffer(
+        spdm_context, (const void **)&cert_chain_buffer, &cert_chain_buffer_size);
+    if (!result) {
+        return false;
+    }
+
+    if (session_info->mut_auth_requested) {
+        result = libspdm_get_local_cert_chain_buffer(
+            spdm_context, (const void **)&mut_cert_chain_buffer,
+            &mut_cert_chain_buffer_size);
+        if (!result) {
+            return false;
+        }
+    } else {
+        mut_cert_chain_buffer = NULL;
+        mut_cert_chain_buffer_size = 0;
+    }
+
+    th_curr_data_size = sizeof(th_curr_data);
+    result = libspdm_calculate_th_for_finish(
+        spdm_context, session_info, cert_chain_buffer,
+        cert_chain_buffer_size, mut_cert_chain_buffer,
+        mut_cert_chain_buffer_size, &th_curr_data_size, th_curr_data);
+    if (!result) {
+        return false;
+    }
+
+    result = libspdm_hash_all (spdm_context->connection_info.algorithm.base_hash_algo,
+                               th_curr_data, th_curr_data_size, hash_data);
+    if (!result) {
+        return false;
+    }
+
+    result = libspdm_hmac_all_with_request_finished_key(
+        session_info->secured_message_context, hash_data,
+        hash_size, calc_hmac_data);
+    if (!result) {
+        return false;
+    }
+#else
+    result = libspdm_calculate_th_hmac_for_finish_req(
+        spdm_context, session_info, &hash_size, calc_hmac_data);
+    if (!result) {
+        return false;
+    }
+#endif
+    LIBSPDM_DEBUG((LIBSPDM_DEBUG_INFO, "th_curr hmac - "));
+    libspdm_internal_dump_data(calc_hmac_data, hash_size);
+    LIBSPDM_DEBUG((LIBSPDM_DEBUG_INFO, "\n"));
+
+    libspdm_copy_mem(hmac, hash_size, calc_hmac_data, hash_size);
+
+    return true;
+}
+
+/**
+ * This function generates the finish signature based upon TH.
+ *
+ * @param  spdm_context                  A pointer to the SPDM context.
+ * @param  session_info                  The session info of an SPDM session.
+ * @param  signature                    The buffer to store the finish signature.
+ *
+ * @retval true  finish signature is generated.
+ * @retval false finish signature is not generated.
+ **/
+static bool libspdm_generate_finish_req_signature(libspdm_context_t *spdm_context,
+                                                  libspdm_session_info_t *session_info,
+                                                  uint8_t *signature)
+{
+    uint8_t hash_data[LIBSPDM_MAX_HASH_SIZE];
+    bool result;
+    size_t signature_size;
+    size_t hash_size;
+#if LIBSPDM_RECORD_TRANSCRIPT_DATA_SUPPORT
+    uint8_t *cert_chain_buffer;
+    size_t cert_chain_buffer_size;
+    uint8_t *mut_cert_chain_buffer;
+    size_t mut_cert_chain_buffer_size;
+    uint8_t th_curr_data[LIBSPDM_MAX_MESSAGE_BUFFER_SIZE];
+    size_t th_curr_data_size;
+#endif
+
+    signature_size = libspdm_get_req_asym_signature_size(
+        spdm_context->connection_info.algorithm.req_base_asym_alg);
+    hash_size = libspdm_get_hash_size(
+        spdm_context->connection_info.algorithm.base_hash_algo);
+
+#if LIBSPDM_RECORD_TRANSCRIPT_DATA_SUPPORT
+    result = libspdm_get_peer_cert_chain_buffer(
+        spdm_context, (const void **)&cert_chain_buffer, &cert_chain_buffer_size);
+    if (!result) {
+        return false;
+    }
+
+    result = libspdm_get_local_cert_chain_buffer(spdm_context,
+                                                 (const void **)&mut_cert_chain_buffer,
+                                                 &mut_cert_chain_buffer_size);
+    if (!result) {
+        return false;
+    }
+
+    th_curr_data_size = sizeof(th_curr_data);
+    result = libspdm_calculate_th_for_finish(
+        spdm_context, session_info, cert_chain_buffer,
+        cert_chain_buffer_size, mut_cert_chain_buffer,
+        mut_cert_chain_buffer_size, &th_curr_data_size, th_curr_data);
+    if (!result) {
+        return false;
+    }
+
+    /* Debug code only - required for debug print of th_curr below*/
+    LIBSPDM_DEBUG_CODE(
+        if (!libspdm_hash_all(
+                spdm_context->connection_info.algorithm.base_hash_algo,
+                th_curr_data, th_curr_data_size, hash_data)) {
+        return false;
+    }
+        );
+#else
+    result = libspdm_calculate_th_hash_for_finish(
+        spdm_context, session_info, &hash_size, hash_data);
+    if (!result) {
+        return false;
+    }
+#endif
+    LIBSPDM_DEBUG((LIBSPDM_DEBUG_INFO, "th_curr hash - "));
+    libspdm_internal_dump_data(hash_data, hash_size);
+    LIBSPDM_DEBUG((LIBSPDM_DEBUG_INFO, "\n"));
+
+#if LIBSPDM_RECORD_TRANSCRIPT_DATA_SUPPORT
+    result = libspdm_requester_data_sign(
+        spdm_context->connection_info.version, SPDM_FINISH,
+        spdm_context->connection_info.algorithm.req_base_asym_alg,
+        spdm_context->connection_info.algorithm.base_hash_algo,
+        false, th_curr_data, th_curr_data_size, signature, &signature_size);
+#else
+    result = libspdm_requester_data_sign(
+        spdm_context->connection_info.version, SPDM_FINISH,
+        spdm_context->connection_info.algorithm.req_base_asym_alg,
+        spdm_context->connection_info.algorithm.base_hash_algo,
+        true, hash_data, hash_size, signature, &signature_size);
+#endif
+    if (result) {
+        LIBSPDM_DEBUG((LIBSPDM_DEBUG_INFO, "signature - "));
+        libspdm_internal_dump_data(signature, signature_size);
+        LIBSPDM_DEBUG((LIBSPDM_DEBUG_INFO, "\n"));
+    }
+
+    return result;
+}
+
+/**
  * This function sends FINISH and receives FINISH_RSP for SPDM finish.
  *
  * @param  spdm_context       A pointer to the SPDM context.
