+++ library/spdm_requester_lib/libspdm_req_get_measurements.c
@@ -24,6 +24,137 @@ typedef struct {
 #pragma pack()
 
 /**
+ * This function verifies the measurement signature based upon l1l2.
+ * If session_info is NULL, this function will use M cache of SPDM context,
+ * else will use M cache of SPDM session context.
+ *
+ * @param  spdm_context                  A pointer to the SPDM context.
+ * @param  session_info                  A pointer to the SPDM session context.
+ * @param  sign_data                     The signature data buffer.
+ * @param  sign_data_size                 size in bytes of the signature data buffer.
+ *
+ * @retval true  signature verification pass.
+ * @retval false signature verification fail.
+ **/
+static bool libspdm_verify_measurement_signature(libspdm_context_t *spdm_context,
+                                                 libspdm_session_info_t *session_info,
+                                                 const void *sign_data,
+                                                 size_t sign_data_size)
+{
+    bool result;
+    const uint8_t *cert_buffer;
+    size_t cert_buffer_size;
+    void *context;
+    const uint8_t *cert_chain_data;
+    size_t cert_chain_data_size;
+#if LIBSPDM_RECORD_TRANSCRIPT_DATA_SUPPORT
+    libspdm_large_managed_buffer_t l1l2;
+    uint8_t *l1l2_buffer;
+    size_t l1l2_buffer_size;
+#else
+    uint8_t l1l2_hash[LIBSPDM_MAX_HASH_SIZE];
+    size_t l1l2_hash_size;
+    uint8_t slot_id;
+#endif
+
+#if LIBSPDM_RECORD_TRANSCRIPT_DATA_SUPPORT
+    result = libspdm_calculate_l1l2(spdm_context, session_info, &l1l2);
+#else
+    l1l2_hash_size = sizeof(l1l2_hash);
+    result = libspdm_calculate_l1l2_hash(spdm_context, session_info, &l1l2_hash_size, l1l2_hash);
+#endif
+    libspdm_reset_message_m(spdm_context, session_info);
+    if (!result) {
+        return false;
+    }
+
+#if LIBSPDM_RECORD_TRANSCRIPT_DATA_SUPPORT
+    l1l2_buffer = libspdm_get_managed_buffer(&l1l2);
+    l1l2_buffer_size = libspdm_get_managed_buffer_size(&l1l2);
+
+    result = libspdm_get_peer_cert_chain_data(
+        spdm_context, (const void **)&cert_chain_data, &cert_chain_data_size);
+    if (!result) {
+        return false;
+    }
+
+    /* Get leaf cert from cert chain*/
+    result = libspdm_x509_get_cert_from_cert_chain(cert_chain_data,
+                                                   cert_chain_data_size, -1,
+                                                   &cert_buffer, &cert_buffer_size);
+    if (!result) {
+        return false;
+    }
+
+    result = libspdm_asym_get_public_key_from_x509(
+        spdm_context->connection_info.algorithm.base_asym_algo,
+        cert_buffer, cert_buffer_size, &context);
+    if (!result) {
+        return false;
+    }
+
+    result = libspdm_asym_verify(
+        spdm_context->connection_info.version, SPDM_MEASUREMENTS,
+        spdm_context->connection_info.algorithm.base_asym_algo,
+        spdm_context->connection_info.algorithm.base_hash_algo, context,
+        l1l2_buffer, l1l2_buffer_size, sign_data, sign_data_size);
+    libspdm_asym_free(spdm_context->connection_info.algorithm.base_asym_algo, context);
+#else
+    slot_id = spdm_context->connection_info.peer_used_cert_chain_slot_id;
+    if (spdm_context->connection_info.peer_used_cert_chain[slot_id].leaf_cert_public_key != NULL) {
+        result = libspdm_asym_verify_hash(
+            spdm_context->connection_info.version, SPDM_MEASUREMENTS,
+            spdm_context->connection_info.algorithm.base_asym_algo,
+            spdm_context->connection_info.algorithm.base_hash_algo,
+            spdm_context->connection_info.peer_used_cert_chain[slot_id].leaf_cert_public_key,
+            l1l2_hash, l1l2_hash_size, sign_data, sign_data_size);
+        if (!result) {
+            LIBSPDM_DEBUG((LIBSPDM_DEBUG_INFO, "!!! verify_measurement_signature - FAIL !!!\n"));
+            return false;
+        }
+
+        LIBSPDM_DEBUG((LIBSPDM_DEBUG_INFO, "!!! verify_measurement_signature - PASS !!!\n"));
+        return true;
+    }
+
+    result = libspdm_get_peer_cert_chain_data(
+        spdm_context, (const void **)&cert_chain_data, &cert_chain_data_size);
+    if (!result) {
+        return false;
+    }
+
+    /* Get leaf cert from cert chain */
+    result = libspdm_x509_get_cert_from_cert_chain(cert_chain_data,
+                                                   cert_chain_data_size, -1,
+                                                   &cert_buffer, &cert_buffer_size);
+    if (!result) {
+        return false;
+    }
+
+    result = libspdm_asym_get_public_key_from_x509(
+        spdm_context->connection_info.algorithm.base_asym_algo,
+        cert_buffer, cert_buffer_size, &context);
+    if (!result) {
+        return false;
+    }
+
+    result = libspdm_asym_verify_hash(
+        spdm_context->connection_info.version, SPDM_MEASUREMENTS,
+        spdm_context->connection_info.algorithm.base_asym_algo,
+        spdm_context->connection_info.algorithm.base_hash_algo, context,
+        l1l2_hash, l1l2_hash_size, sign_data, sign_data_size);
+    libspdm_asym_free(spdm_context->connection_info.algorithm.base_asym_algo, context);
+#endif
+    if (!result) {
+        LIBSPDM_DEBUG((LIBSPDM_DEBUG_INFO, "!!! verify_measurement_signature - FAIL !!!\n"));
+        return false;
+    }
+
+    LIBSPDM_DEBUG((LIBSPDM_DEBUG_INFO, "!!! verify_measurement_signature - PASS !!!\n"));
+    return true;
+}
+
+/**
  * This function sends GET_MEASUREMENT to get measurement from the device.
  * If the signature is requested this function verifies the signature of the measurement.
  *
