+++ library/spdm_responder_lib/libspdm_rsp_psk_exchange.c
@@ -9,6 +9,68 @@
 #if LIBSPDM_ENABLE_CAPABILITY_PSK_EX_CAP
 
 /**
+ * This function generates the PSK exchange HMAC based upon TH.
+ *
+ * @param  spdm_context                  A pointer to the SPDM context.
+ * @param  session_info                  The session info of an SPDM session.
+ * @param  hmac                         The buffer to store the PSK exchange HMAC.
+ *
+ * @retval true  PSK exchange HMAC is generated.
+ * @retval false PSK exchange HMAC is not generated.
+ **/
+static bool libspdm_generate_psk_exchange_rsp_hmac(libspdm_context_t *spdm_context,
+                                                   libspdm_session_info_t *session_info,
+                                                   uint8_t *hmac)
+{
+    uint8_t hmac_data[LIBSPDM_MAX_HASH_SIZE];
+    size_t hash_size;
+    bool result;
+#if LIBSPDM_RECORD_TRANSCRIPT_DATA_SUPPORT
+    uint8_t th_curr_data[LIBSPDM_MAX_MESSAGE_BUFFER_SIZE];
+    size_t th_curr_data_size;
+    uint8_t hash_data[LIBSPDM_MAX_HASH_SIZE];
+#endif
+
+    hash_size = libspdm_get_hash_size(spdm_context->connection_info.algorithm.base_hash_algo);
+
+#if LIBSPDM_RECORD_TRANSCRIPT_DATA_SUPPORT
+    th_curr_data_size = sizeof(th_curr_data);
+    result = libspdm_calculate_th_for_exchange(spdm_context, session_info,
+                                               NULL, 0, &th_curr_data_size,
+                                               th_curr_data);
+    if (!result) {
+        return false;
+    }
+
+    result = libspdm_hash_all (spdm_context->connection_info.algorithm.base_hash_algo,
+                               th_curr_data, th_curr_data_size, hash_data);
+    if (!result) {
+        return false;
+    }
+
+    result = libspdm_hmac_all_with_response_finished_key(
+        session_info->secured_message_context, hash_data,
+        hash_size, hmac_data);
+    if (!result) {
+        return false;
+    }
+#else
+    result = libspdm_calculate_th_hmac_for_exchange_rsp(
+        spdm_context, session_info, false, &hash_size, hmac_data);
+    if (!result) {
+        return false;
+    }
+#endif
+    LIBSPDM_DEBUG((LIBSPDM_DEBUG_INFO, "th_curr hmac - "));
+    libspdm_internal_dump_data(hmac_data, hash_size);
+    LIBSPDM_DEBUG((LIBSPDM_DEBUG_INFO, "\n"));
+
+    libspdm_copy_mem(hmac, hash_size, hmac_data, hash_size);
+
+    return true;
+}
+
+/**
  * Process the SPDM PSK_EXCHANGE request and return the response.
  *
  * @param  spdm_context                  A pointer to the SPDM context.
