+++ library/spdm_responder_lib/libspdm_rsp_psk_finish.c
@@ -7,6 +7,75 @@
 #include "internal/libspdm_responder_lib.h"
 #include "hal/library/platform_lib.h"
 
+#if LIBSPDM_ENABLE_CAPABILITY_PSK_EX_CAP
+
+/**
+ * This function verifies the PSK finish HMAC based upon TH.
+ *
+ * @param  spdm_context                  A pointer to the SPDM context.
+ * @param  session_info                  The session info of an SPDM session.
+ * @param  hmac_data                     The HMAC data buffer.
+ * @param  hmac_data_size                 size in bytes of the HMAC data buffer.
+ *
+ * @retval true  HMAC verification pass.
+ * @retval false HMAC verification fail.
+ **/
+bool libspdm_verify_psk_finish_req_hmac(libspdm_context_t *spdm_context,
+                                        libspdm_session_info_t *session_info,
+                                        const uint8_t *hmac, size_t hmac_size)
+{
+    uint8_t hmac_data[LIBSPDM_MAX_HASH_SIZE];
+    size_t hash_size;
+    bool result;
+#if LIBSPDM_RECORD_TRANSCRIPT_DATA_SUPPORT
+    uint8_t th_curr_data[LIBSPDM_MAX_MESSAGE_BUFFER_SIZE];
+    size_t th_curr_data_size;
+    uint8_t hash_data[LIBSPDM_MAX_HASH_SIZE];
+#endif
+
+    hash_size = libspdm_get_hash_size(spdm_context->connection_info.algorithm.base_hash_algo);
+    LIBSPDM_ASSERT(hmac_size == hash_size);
+
+#if LIBSPDM_RECORD_TRANSCRIPT_DATA_SUPPORT
+    th_curr_data_size = sizeof(th_curr_data);
+    result = libspdm_calculate_th_for_finish(spdm_context, session_info, NULL,
+                                             0, NULL, 0, &th_curr_data_size,
+                                             th_curr_data);
+    if (!result) {
+        return false;
+    }
+
+    result = libspdm_hash_all (spdm_context->connection_info.algorithm.base_hash_algo,
+                               th_curr_data, th_curr_data_size, hash_data);
+    if (!result) {
+        return false;
+    }
+
+    result = libspdm_hmac_all_with_request_finished_key(
+        session_info->secured_message_context, hash_data,
+        hash_size, hmac_data);
+    if (!result) {
+        return false;
+    }
+#else
+    result = libspdm_calculate_th_hmac_for_finish_req(
+        spdm_context, session_info, &hash_size, hmac_data);
+    if (!result) {
+        return false;
+    }
+#endif
+    LIBSPDM_DEBUG((LIBSPDM_DEBUG_INFO, "Calc th_curr hmac - "));
+    libspdm_internal_dump_data(hmac_data, hash_size);
+    LIBSPDM_DEBUG((LIBSPDM_DEBUG_INFO, "\n"));
+
+    if (libspdm_const_compare_mem(hmac, hmac_data, hash_size) != 0) {
+        LIBSPDM_DEBUG((LIBSPDM_DEBUG_INFO, "!!! verify_psk_finish_req_hmac - FAIL !!!\n"));
+        return false;
+    }
+    LIBSPDM_DEBUG((LIBSPDM_DEBUG_INFO, "!!! verify_psk_finish_req_hmac - PASS !!!\n"));
+    return true;
+}
+
 /**
  * Process the SPDM PSK_FINISH request and return the response.
  *
@@ -185,3 +254,5 @@ libspdm_return_t libspdm_get_response_psk_finish(void *context,
 
     return LIBSPDM_STATUS_SUCCESS;
 }
+
+#endif /* LIBSPDM_ENABLE_CAPABILITY_PSK_EX_CAP */
